Deployment/requirements.txt
streamlit
pandas
numpy
scikit.learn
tensorflow
tensorflow_hub
Pillow